Move workaround into swagger config
