+++ src/main/java/se/gewalli/config/SwaggerConfig.java
@@ -1,11 +1,30 @@
 package se.gewalli.config;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import javax.servlet.ServletContext;
+
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
+import org.springframework.web.servlet.mvc.method.RequestMappingInfo.BuilderConfiguration;
+import org.springframework.web.servlet.mvc.method.RequestMappingInfoHandlerMapping;
+
+import springfox.documentation.RequestHandler;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spi.service.RequestHandlerProvider;
+import springfox.documentation.spring.web.WebMvcRequestHandler;
+import springfox.documentation.spring.web.paths.Paths;
 import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.spring.web.plugins.DocumentationPluginsBootstrapper;
+import springfox.documentation.spring.web.plugins.WebMvcRequestHandlerProvider;
+import springfox.documentation.spring.web.readers.operation.HandlerMethodResolver;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @Configuration
@@ -18,5 +37,30 @@ public class SwaggerConfig {
                 .apis(RequestHandlerSelectors.any())
                 .paths(PathSelectors.any())
                 .build();
+    }
+    @Bean
+    public InitializingBean removeSpringfoxHandlerProvider(DocumentationPluginsBootstrapper bootstrapper) {
+        return () -> bootstrapper.getHandlerProviders().removeIf(WebMvcRequestHandlerProvider.class::isInstance);
+    }
+    /** 
+     * Note removeSpringfoxHandlerProvider, this workaround is described here: 
+     * https://github.com/springfox/springfox/issues/3462#issuecomment-1076552144
+    */
+    @Bean
+    public RequestHandlerProvider customRequestHandlerProvider(Optional<ServletContext> servletContext, HandlerMethodResolver methodResolver, List<RequestMappingInfoHandlerMapping> handlerMappings) {
+        String contextPath = servletContext.map(ServletContext::getContextPath).orElse(Paths.ROOT);
+        return () -> handlerMappings.stream()
+            .filter(mapping -> !mapping.getClass().getSimpleName().equals("IntegrationRequestMappingHandlerMapping"))
+            .map(mapping -> mapping.getHandlerMethods().entrySet())
+            .flatMap(Set::stream)
+            .map(entry -> new WebMvcRequestHandler(contextPath, methodResolver, tweakInfo(entry.getKey()), entry.getValue()))
+            .sorted(RequestHandler.byPatternsCondition())
+            .collect(Collectors.toList());
+    }
+
+    RequestMappingInfo tweakInfo(RequestMappingInfo info) {
+        if (info.getPathPatternsCondition() == null) return info;
+        String[] patterns = info.getPathPatternsCondition().getPatternValues().toArray(String[]::new);
+        return info.mutate().options(new BuilderConfiguration()).paths(patterns).build();
     }
 }
